pyproject.toml
[project]
name = "googlecampus-midterm"
version = "0.1.0"
description = "Gemini AI Demo"
authors = [
    {name = "Life-00",email = "kanda0101@naver.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "google-cloud-aiplatform (>=1.124.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
